remove old popups
