+++ src/components/ui/popups/common/AccessPolygonMap.svelte
