+++ src/components/ui/popups/common/BasicPokemonDisplayMultiple.svelte
