+++ src/components/ui/popups/common/BasicPokemonDisplayOne.svelte
