+++ src/components/ui/popups/common/BigExpireTime.svelte
