+++ src/components/ui/popups/common/InvasionLineupEntry.svelte
