+++ src/components/ui/popups/common/MainAccessMap.svelte
