+++ src/components/ui/popups/common/MainCardBigIcon.svelte
