+++ src/components/ui/popups/common/PopupBaseStatic.svelte
