+++ src/components/ui/popups/common/StatsMainCardEntry.svelte
