+++ src/components/ui/popups/pokemon/IvBreakdown.svelte
