+++ src/components/ui/popups/tappable/TappablePopup.svelte
@@ -1,73 +1,74 @@
-<script lang="ts">
-	import { getMapObjects } from "@/lib/mapObjects/mapObjectsState.svelte";
-	import {
-		getCurrentSelectedData,
-		getCurrentSelectedMapId
-	} from "@/lib/mapObjects/currentSelectedState.svelte";
-	import BasePopup from "@/components/ui/popups/BasePopup.svelte";
+<script module lang="ts">
+	import type { MapObjectPopupProps } from "@/components/ui/popups/common/PopupBaseStatic.svelte";
+	import * as m from "$lib/paraglide/messages";
+	import { mPokemon } from "$lib/services/ingameLocale";
+	import type { MapData } from "$lib/mapObjects/mapObjectTypes";
+	import { MapObjectType } from "$lib/mapObjects/mapObjectTypes";
 	import ImagePopup from "@/components/ui/popups/common/ImagePopup.svelte";
-	import type { TappableData } from "@/lib/types/mapObjectData/tappable.d.ts";
-	import { mItem, mPokemon } from "@/lib/services/ingameLocale";
-	import { getIconPokemon, getIconTappable } from "@/lib/services/uicons.svelte";
-	import { m } from "@/lib/paraglide/messages";
-	import { Clock, ClockAlert, RotateCcw, Trees } from "@lucide/svelte";
-	import { formatDecimal } from "@/lib/utils/numberFormat";
-	import { hasTimer } from "@/lib/utils/pokemonUtils";
-	import IconValue from "@/components/ui/popups/common/IconValue.svelte";
-	import TimeWithCountdown from "@/components/ui/popups/common/TimeWithCountdown.svelte";
+	import OverviewCard from "@/components/ui/popups/common/OverviewCard.svelte";
+	import TitledMainSection from "@/components/ui/popups/common/TitledMainSection.svelte";
+	import StatsMainCard from "@/components/ui/popups/common/StatsMainCard.svelte";
+	import StatsMainCardEntry from "@/components/ui/popups/common/StatsMainCardEntry.svelte";
+	import UpdatedTimes from "@/components/ui/popups/common/UpdatedTimes.svelte";
+	import { getIconPokemon, getIconPokestop, getIconTappable } from "$lib/services/uicons.svelte";
+	import { formatNumber, formatPercentage, formatDecimal } from "$lib/utils/numberFormat";
+	import { CircleDot, CircleSlash2, Clock, Info, MapPinned, RotateCcw, Trees, VectorSquare } from "@lucide/svelte";
+	import type { NestData } from "$lib/types/mapObjectData/nest";
+	import AccessPolygonMap from "@/components/ui/popups/common/AccessPolygonMap.svelte";
+	import type { TappableData } from "$lib/types/mapObjectData/tappable.d.ts";
+	import { getTappableName } from "$lib/utils/tappableUtils";
+	import { setActiveSearchPokemon } from "$lib/features/activeSearch.svelte";
+	import QuickSearchButton from "@/components/ui/popups/common/QuickSearchButton.svelte";
+	import { hasTimer } from "$lib/utils/pokemonUtils";
+	import { timestampToLocalTime } from "$lib/utils/timestampToLocalTime";
 	import Countdown from "@/components/utils/Countdown.svelte";
-	import { getTappableName } from "@/lib/utils/tappableUtils";
+	import BasicMainCard from "@/components/ui/popups/common/BasicMainCard.svelte";
+	import { resize } from "$lib/services/assets";
+	import MainAccessMap from "@/components/ui/popups/common/MainAccessMap.svelte";
 
-	let data: TappableData = $derived(
-		(getMapObjects()[getCurrentSelectedMapId()] as TappableData) ??
-			(getCurrentSelectedData() as TappableData)
-	);
+	export { image, main };
+
+	export function getPopupPropsTappable(data: MapData) {
+		data = data as TappableData;
+		return {
+			type: m.pogo_tappable(),
+			title: getTappableName(data),
+			image,
+			main
+		} as MapObjectPopupProps;
+	}
+</script>
+<script>
+	import BigCountdown from "@/components/ui/popups/common/BigCountdown.svelte";
 </script>
 
-{#snippet basicInfo()}
-	<IconValue Icon={Clock}>
-		<span>
-			{m.popup_despawns()}
-		</span>
-
-		<TimeWithCountdown expireTime={data.expire_timestamp} />
-	</IconValue>
-	{#if !hasTimer(data)}
-		<IconValue Icon={ClockAlert}>
-			{m.time_is_estimated()}
-		</IconValue>
-	{/if}
+{#snippet image(d: MapData)}
+	{@const data = d as TappableData}
+	<div class="size-14 shrink-0">
+		<ImagePopup alt={getTappableName(data)} src={getIconTappable(data)} class="w-12 h-12" />
+	</div>
 {/snippet}
 
-<BasePopup lat={data.lat} lon={data.lon}>
-	{#snippet image()}
-		<div class="w-12 shrink-0">
-			<ImagePopup alt={getTappableName(data)} src={getIconTappable(data)} class="w-12 h-12" />
-		</div>
-	{/snippet}
+{#snippet main(d: MapData)}
+	{@const data = d as TappableData}
 
-	{#snippet title()}
-		<div class="text-lg font-semibold tracking-tight">
-			<span>
-				{getTappableName(data)} ({m.pogo_tappable()})
-			</span>
-		</div>
-	{/snippet}
+	<BigCountdown
+		expire={data?.expire_timestamp ?? 0}
+		fallbackExpire={data.updated}
+		useFallback={!hasTimer(data)}
+		fallbackTitle={m.last_seen()}
+		fallbackExplanation={m.unknown_spawnpoint_notice()}
+	/>
 
-	{#snippet description()}
-		{@render basicInfo()}
-	{/snippet}
-
-	{#snippet content()}
-		<div class="mb-3">
-			{@render basicInfo()}
-		</div>
-
-		<IconValue Icon={Clock}>
-			{m.last_updated()}:
-			<b>
-				<Countdown expireTime={data.updated} />
-			</b>
-		</IconValue>
-	{/snippet}
-</BasePopup>
+	<TitledMainSection Icon={CircleDot} title={m.access_this_tappable()}>
+		<MainAccessMap
+			lat={data.lat}
+			lon={data.lon}
+			type={MapObjectType.TAPPABLE}
+			uiconType="tappable"
+			radius={40}
+			zoom={16.5}
+			icon={resize(getIconTappable(data), { width: 64 })}
+		/>
+	</TitledMainSection>
+{/snippet}
